Users is int for now as we don't have millions of users to justify the overhead, Yet!

+++ server/src/main/java/dev/findfirst/security/jwt/service/RefreshTokenService.java
@@ -47,7 +47,7 @@ public class RefreshTokenService {
   }
 
   @Transactional
-  public int deleteByUserId(Long userId) {
+  public int deleteByUserId(Integer userId) {
     return refreshTokenRepository.deleteByUser(userRepository.findById(userId).get());
   }
 }
